+++ lib/graphql/static_validation/rules/argument_literals_are_compatible.rb
@@ -37,14 +37,19 @@ module GraphQL
                 # short-circuit here so we avoid bubbling this up to whatever input_object / array contains us
                 return super
               end
-              error = GraphQL::StaticValidation::ArgumentLiteralsAreCompatibleError.new(err.message, nodes: parent, type: "CoercionError")
+              error = GraphQL::StaticValidation::ArgumentLiteralsAreCompatibleError.new(err.message, nodes: parent, type: "CoercionError", extensions: err.extensions)
             rescue GraphQL::LiteralValidationError => err
               # check to see if the ast node that caused the error to be raised is
               # the same as the node we were checking here.
-              matched = if arg_defn.type.kind.list?
+              arg_type = arg_defn.type
+              if arg_type.kind.non_null?
+                arg_type = arg_type.of_type
+              end
+
+              matched = if arg_type.kind.list?
                 # for a list we claim an error if the node is contained in our list
                 Array(node.value).include?(err.ast_value)
-              elsif arg_defn.type.kind.input_object? && node.value.is_a?(GraphQL::Language::Nodes::InputObject)
+              elsif arg_type.kind.input_object? && node.value.is_a?(GraphQL::Language::Nodes::InputObject)
                 # for an input object we check the arguments
                 node.value.arguments.include?(err.ast_value)
               else