+++ lib/graphql/static_validation/rules/argument_literals_are_compatible_error.rb
@@ -5,10 +5,11 @@ module GraphQL
       attr_reader :type_name
       attr_reader :argument_name
 
-      def initialize(message, path: nil, nodes: [], type:, argument: nil)
+      def initialize(message, path: nil, nodes: [], type:, argument: nil, extensions: nil)
         super(message, path: path, nodes: nodes)
         @type_name = type
         @argument_name = argument
+        @extensions = extensions
       end
 
       # A hash representation of this Message
@@ -17,7 +18,7 @@ module GraphQL
           "code" => code,
           "typeName" => type_name
         }.tap { |h| h["argumentName"] = argument_name unless argument_name.nil? }
-
+        extensions.merge!(@extensions) unless @extensions.nil?
         super.merge({
           "extensions" => extensions
         })