+++ spec/graphql/subscriptions_spec.rb
@@ -459,7 +459,7 @@ describe GraphQL::Subscriptions do
           }
             GRAPHQL
 
-            res = schema.execute(query_str, context: { socket: "1", me: "1" }, variables: { "type" => "ONE" }, root_value: root_object)
+            schema.execute(query_str, context: { socket: "1", me: "1" }, variables: { "type" => "ONE" }, root_value: root_object)
             err = assert_raises(RuntimeError) {
               schema.subscriptions.trigger("myEvent", { "type" => "ONE" }, error_payload_class.new, scope: "1")
             }